removed defunced docs/ and added doc/qraat-doc.tex
This is a part of GNU Radio documentation system that we're not using
for this project. If we do plan to hook our work into GNU Radio somewhere
down the line, we'll want to rebase our code on the example processing
block provided in the GR distribution. Also, added a .tex and .pdf
documenting the configuration that's needed for the protoype QRAAT
implementation.