Skeleton of qraat.est.
We've decided to eliminate the functionality of est_data/data_arrays.
Our needs now are a quick way to append est's to files by tagname and
to output directly to database. We could implement this in qraat.det,
but we will need to insert csv files into the database; it would be
useful if these were handled by the same interface. Hence, qraat.est!