Began qraat.est.read_db().
The current header for pulse records has only the name of
the transmitter. We can only infer the source site by the
directory structure of the stored det files. The first
versioned metadata header will have (tx_id, site_id), serial
id's in the DB. read_db() and write_db() need to be written
to handle the legacy and versioned metadata.