Changed map page, fixed timezone issues
The map is now able to be color coded by time, and a portion of time can
be selected in the flot plot, which is also colored. This is done using
bitmap images because rendering a thousand vector circles takes too
long. (Although, once we add in point reduction, we might never display
that many points at once)
You can also select time with a slider, although this is only applicable
for the heatmap and breaks down if you move the slider too fast.
Updating the map is now done using Ajax.
The form sidebar is slightly updated: Portions are now collapsable to
save space, and new controls were added.
Timezone issues were fixed by adding a file containing TIMEZONE =
"America/Los_Angeles", and using sym links to use it in both python and
to {% include %} it in the base.html. (And of course changing functions
to use a specific timezone instead of local time or -7*60*60.) (These
new date/time functions reside in utils.py)