Fixed updating data, introduced time slider, and selecting the map
Updating data should provide correct data, and the latlngData array for
the polyline and heatmap should also update correctly. latlngData only
updates when polyline/heatmap is displayed.

The time slider works as intended, but is laggy when using Firefox. I
believe it might be the fact that drawing the flot plot (drawing an arc
for every point) takes 10 times as long in Firefox than in Chrome. So
sliding the slider, instead of clicking it, is pretty bad.
But otherwise, the slider is a working alternative to selecting the flot
plot, and is a cool way to gradually see points in time.

You can now select points on the map by shift clicking and dragging a
rectangle. It will highlight the selected time period in the flot plot
in green. It only selects points in the selected time frame, and will
update if the time selection is changed (by the slider/flot plot).