Semi-functional slider for filters
There's now a slider on the left of the flot plot that you can adjust
the likelihood/actiity filters with. When you switch between the two,
the state of the slider is preserved and reapplied when you switch back.

Adjusting the filters like this also updates the polylines and heatmap.
The polyline is not segmented; It just draws a line through all the
points given (which are ordered by time), regardless of how far apart in
time they are.