Filter slider and form inpus now synced.
Sliders start with the values in the form. And now changing one changes
the other.

Accounting for the interactive filters while changing the time frame is
a pain with my current method. I currently determine the start and
ending index of the current time frame within rawPosData and
remove/insert/push/pop accordingly. To deal with the filters, I would
have to put an ugly check everywhere I push/pop.

I may change it so that it just creates a new array each time the time
selection is changed.