Scaffolding for multiple deployments and flot plots
Can kind of display multiple deployments on the map by specifying in the
URL. IE, /project/#/deployment/#+#+.../

I created extra divs so we have a flot plot per dep, and changed
references where necessary. But it turns out that we want multiple
deployments on one plot, and one set of filters to apply to all the
deployments.

So some changes will have to be undone.