Telemetry page has safer datetime input
Form only shows one pair of date time fields at once, such that you
cannot input conflicting information.
There is also now a "canonical" time format (timestamp) that is to be
passed into the url. "start_timestamp" and "end_timestamp" must be provided.
(The purpose of chosing this is to avoid daylight
savings problems). If the user constructs a bad URL, the existing "if
hiearchy" in the view will take care of that, and the user will also
get an alert on the page warning them of a bad URL that might contain
conflicting information.